refactor: Introduce ChasseurBuilder for improved test object creation
